pyproject.toml
[project]
name = "generic-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.1",
    "langchain>=1.0.1",
    "ollama>=0.6.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]
